# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GroChain Backend API",
    "description": "Comprehensive agricultural supply chain management platform API",
    "version": "1.0.0",
    "contact": {
      "name": "GroChain Support",
      "email": "support@grochain.ng"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.grochain.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Harvest": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "farmer": { "type": "string" },
          "cropType": { "type": "string" },
          "quantity": { "type": "number" },
          "unit": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "verified", "approved", "rejected"] },
          "batchId": { "type": "string" },
          "location": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/harvests": {
      "get": {
        "summary": "Get harvests",
        "description": "Retrieve harvests with optional filtering",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": { "type": "string", "enum": ["pending", "verified", "approved", "rejected"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Harvests retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Harvest" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create harvest",
        "description": "Create a new harvest record",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["cropType", "quantity", "unit", "location"],
                "properties": {
                  "cropType": { "type": "string" },
                  "quantity": { "type": "number" },
                  "unit": { "type": "string" },
                  "location": { "type": "string" },
                  "description": { "type": "string" },
                  "images": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Harvest created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Harvest" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}